requirements.txt
dataclasses==0.5
dj-database-url==0.5.0
Django==2.0.5
django-heroku==0.3.1
djangorestframework==3.8.2
djangotoolbox==1.8.0
djongo==1.2.28
gunicorn==19.8.1
psycopg2==2.7.4
pymongo==3.6.1
pytz==2018.4
sqlparse==0.2.4
whitenoise==3.3.1
